feat(cordovaserve): support option to use Dotenv plugin (#1)
This supports the upgrade in mobile to Angular 11. If you recall, we patched 
this when upgrading to Angular 10 to maintain support for our use of Dotenv 
with the Cordova serve builder.

As part of upgrading this to the latest version, with support for Angular 11, 
I also improved the Dotenv plugin support by using the builder options to 
pass a path to the options in, rather than hardcoding the options into this patch.